requirements.txt

# Streamlit
streamlit>=1.45.0
watchdog>=4.0.0
# ── Core data stack ───────────────────────────────────────────────────────────
numpy>=1.24,<2.0
pandas>=2.0,<3.0
pyarrow>=14.0

# ── Machine learning ──────────────────────────────────────────────────────────
scikit-learn>=1.3,<2.0
xgboost>=2.0,<3.0
joblib>=1.3

# ── Visualisation ─────────────────────────────────────────────────────────────
matplotlib>=3.7,<4.0
seaborn>=0.13
scipy>=1.11

# ── Configuration ─────────────────────────────────────────────────────────────
pyyaml>=6.0

# ── Notebooks ─────────────────────────────────────────────────────────────────
jupyter>=1.0
nbformat>=5.9
ipykernel>=6.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=7.4
pytest-cov>=4.1